# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOGISTIK UNIKOM API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/barang": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "List barang",
                "operationId": "f9b86624b1ea03a5758e8eae04bc6621",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Barang"
                ],
                "summary": "Create barang",
                "operationId": "babddf8dd1e89ef20f9f701a5d0b130e",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang/{id}": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Detail barang",
                "operationId": "57e3d556d5baa0e628ac33da7d19b222",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Barang"
                ],
                "summary": "Update barang",
                "operationId": "a8412c1cad62703ba5a3e430dc6bc5c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Barang"
                ],
                "summary": "Delete barang",
                "operationId": "3bb31845049c03c5405f083825d70c1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-pengajuan": {
            "post": {
                "tags": [
                    "Barang Pengajuan"
                ],
                "summary": "Tambah barang pengajuan",
                "operationId": "6cb519615229bc3c9d175f1032b047db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "barang_id",
                                    "jumlah"
                                ],
                                "properties": {
                                    "barang_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "jumlah": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Untuk kebutuhan operasional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Barang berhasil ditambahkan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-pengajuan/{id}/approve": {
            "patch": {
                "tags": [
                    "Barang Pengajuan"
                ],
                "summary": "Approve barang pengajuan",
                "operationId": "e622cf87a6d82dd68ad95b850f98d203",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Barang Pengajuan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "jumlah_disetujui",
                                    "status"
                                ],
                                "properties": {
                                    "jumlah_disetujui": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status": {
                                        "description": "1=approve, 2=reject",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barang berhasil di approve"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-pengajuan/{id}": {
            "put": {
                "tags": [
                    "Barang Pengajuan"
                ],
                "summary": "Update barang pengajuan",
                "operationId": "d7720190bf54920ef1b4a4d7c8e49240",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Barang Pengajuan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "barang_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "jumlah": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Update kebutuhan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barang berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Barang Pengajuan"
                ],
                "summary": "Hapus barang pengajuan",
                "operationId": "23321121d35768c1af1b39a00480acc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Barang Pengajuan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}